third_party/blink/web_tests/platform/generic/fast/events/tabindex-no-focusable-all-negative-expected.txt

PASS successfullyParsed is true

TEST COMPLETE

PASS document.activeElement.id is "focusMe"
PASS document.activeElement.id is "focusMe"
  
